refactor(auth): 优化用户ID类型和数据库主键类型统一
- 将用户相关实体类主键从UUID字符串改为自增Long类型
- 修改数据库初始化脚本，调整表结构和字段类型，支持自增主键
- 统一用户上下文、拦截器和JWT工具中用户ID的数据类型为Long
- 更新JWT处理逻辑，支持Long类型用户ID的创建和解析
- 修正权限切面中checkRole方法调用签名及参数传递
- 调整相关Mapper接口，使用Long类型作为参数和返回值类型

+++ src/main/java/com/seuoj/seuojbackend/entity/Tag.java
@@ -11,9 +11,6 @@ import java.time.LocalDateTime;
 
 /**
  * 题目标签表实体类
- *
- * @author YourName
- * @since 2025-12-13
  */
 @Data
 @EqualsAndHashCode(callSuper = false)
@@ -29,7 +26,7 @@ public class Tag implements Serializable {
      * 主键ID
      */
     @TableId(value = "id", type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     /**
      * 标签名称
